[CoinMarketCap] Move old deprecated API v2 to separate package
