+++ xchange-coinmarketcap/src/main/java/org/knowm/xchange/coinmarketcap/deprecated/v2/CoinMarketCap.java
@@ -1,13 +1,17 @@
-package org.knowm.xchange.coinmarketcap;
+package org.knowm.xchange.coinmarketcap.deprecated.v2;
 
-import java.io.IOException;
-import java.util.Objects;
-import javax.ws.rs.*;
-import javax.ws.rs.core.MediaType;
-import org.knowm.xchange.coinmarketcap.dto.marketdata.CoinMarketCapArrayData;
-import org.knowm.xchange.coinmarketcap.dto.marketdata.CoinMarketCapTicker;
+import org.knowm.xchange.coinmarketcap.deprecated.v2.dto.marketdata.CoinMarketCapArrayData;
+import org.knowm.xchange.coinmarketcap.deprecated.v2.dto.marketdata.CoinMarketCapTicker;
 import org.knowm.xchange.currency.CurrencyPair;
 import org.knowm.xchange.utils.jackson.CurrencyPairDeserializer;
+
+import javax.ws.rs.GET;
+import javax.ws.rs.Path;
+import javax.ws.rs.Produces;
+import javax.ws.rs.QueryParam;
+import javax.ws.rs.core.MediaType;
+import java.io.IOException;
+import java.util.Objects;
 
 /** @author allenday https://api.coinmarketcap.com/v2/ticker/ */
 @Path("/v2")
