+++ xchange-coinmarketcap/src/main/java/org/knowm/xchange/coinmarketcap/deprecated/v2/CoinMarketCapExchange.java
@@ -1,12 +1,9 @@
-package org.knowm.xchange.coinmarketcap;
+package org.knowm.xchange.coinmarketcap.deprecated.v2;
 
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
 import org.knowm.xchange.BaseExchange;
 import org.knowm.xchange.Exchange;
 import org.knowm.xchange.ExchangeSpecification;
-import org.knowm.xchange.coinmarketcap.service.CoinMarketCapMarketDataService;
+import org.knowm.xchange.coinmarketcap.deprecated.v2.service.CoinMarketCapMarketDataService;
 import org.knowm.xchange.currency.Currency;
 import org.knowm.xchange.currency.CurrencyPair;
 import org.knowm.xchange.dto.meta.ExchangeMetaData;
@@ -16,6 +13,10 @@ import org.knowm.xchange.service.marketdata.MarketDataService;
 import org.knowm.xchange.service.trade.TradeService;
 import org.knowm.xchange.utils.nonce.CurrentTimeNonceFactory;
 import si.mazi.rescu.SynchronizedValueFactory;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
 
 /** @author allenday */
 public class CoinMarketCapExchange extends BaseExchange implements Exchange {
