+++ xchange-coinmarketcap/src/main/java/org/knowm/xchange/coinmarketcap/deprecated/v2/service/CoinMarketCapMarketDataServiceRaw.java
@@ -1,17 +1,18 @@
-package org.knowm.xchange.coinmarketcap.service;
+package org.knowm.xchange.coinmarketcap.deprecated.v2.service;
 
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
 import org.knowm.xchange.Exchange;
-import org.knowm.xchange.coinmarketcap.CoinMarketCap;
-import org.knowm.xchange.coinmarketcap.dto.marketdata.CoinMarketCapCurrency;
-import org.knowm.xchange.coinmarketcap.dto.marketdata.CoinMarketCapTicker;
+import org.knowm.xchange.coinmarketcap.deprecated.v2.CoinMarketCap;
+import org.knowm.xchange.coinmarketcap.deprecated.v2.dto.marketdata.CoinMarketCapCurrency;
+import org.knowm.xchange.coinmarketcap.deprecated.v2.dto.marketdata.CoinMarketCapTicker;
 import org.knowm.xchange.currency.Currency;
 import org.knowm.xchange.currency.CurrencyPair;
 import org.knowm.xchange.service.BaseExchangeService;
 import org.knowm.xchange.service.BaseService;
 import si.mazi.rescu.RestProxyFactory;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
 
 /** @author allenday */
 class CoinMarketCapMarketDataServiceRaw extends BaseExchangeService implements BaseService {
