+++ xchange-coinmarketcap/src/test/java/org/knowm/xchange/coinmarketcap/deprecated/v2/service/marketdata/TickerFetchIntegration.java
@@ -1,20 +1,21 @@
-package org.knowm.xchange.coinmarketcap.service.marketdata;
+package org.knowm.xchange.coinmarketcap.deprecated.v2.service.marketdata;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.knowm.xchange.currency.CurrencyPair.NEO_USD;
-
-import java.util.List;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.knowm.xchange.Exchange;
 import org.knowm.xchange.ExchangeFactory;
-import org.knowm.xchange.coinmarketcap.CoinMarketCapExchange;
-import org.knowm.xchange.coinmarketcap.dto.marketdata.CoinMarketCapTicker;
-import org.knowm.xchange.coinmarketcap.service.CoinMarketCapMarketDataService;
+import org.knowm.xchange.coinmarketcap.deprecated.v2.CoinMarketCapExchange;
+import org.knowm.xchange.coinmarketcap.deprecated.v2.dto.marketdata.CoinMarketCapTicker;
+import org.knowm.xchange.coinmarketcap.deprecated.v2.service.CoinMarketCapMarketDataService;
 import org.knowm.xchange.currency.Currency;
 import org.knowm.xchange.currency.CurrencyPair;
 import org.knowm.xchange.dto.marketdata.Ticker;
 import org.knowm.xchange.service.marketdata.MarketDataService;
+
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.knowm.xchange.currency.CurrencyPair.NEO_USD;
 
 /** @author allenday */
 public class TickerFetchIntegration {
