+++ xchange-coinmarketcap/src/test/resources/org/knowm/xchange/coinmarketcap/deprecated/v2/dto/marketdata/example-ticker-data-v2.json
